added dao tests

+++ QuizWebsite/src/main/java/org/ja/dao/AnnouncementsDao.java
@@ -3,6 +3,7 @@ package org.ja.dao;
 import org.apache.commons.dbcp2.BasicDataSource;
 import org.ja.model.OtherObjects.Announcement;
 import org.ja.model.OtherObjects.History;
+import org.ja.model.OtherObjects.QuizRating;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
@@ -23,13 +24,16 @@ create table announcements(
 
 public class AnnouncementsDao {
     private final BasicDataSource dataSource;
-
+    private long cnt=0;
     public AnnouncementsDao(BasicDataSource dataSource) {
         this.dataSource = dataSource;
     }
 
     public void insertAnnouncement(Announcement announcement){
-        String sql = "INSERT into history (administrator_id, announcement_text) VALUES (?, ?)";
+        if(contains(announcement)){
+            return;
+        }
+        String sql = "INSERT into announcements (administrator_id, announcement_text) VALUES (?, ?)";
 
         try (Connection c = dataSource.getConnection();
              PreparedStatement ps = c.prepareStatement(sql, PreparedStatement.RETURN_GENERATED_KEYS)){
@@ -40,8 +44,11 @@ public class AnnouncementsDao {
             ps.executeUpdate();
 
             try (ResultSet rs = ps.getGeneratedKeys()){
-                if (rs.next())
+                if (rs.next()){
+                    cnt++;
+                    announcement.setCreationDate(rs.getTimestamp("creation_date"));
                     announcement.setAnnouncementId(rs.getLong(1));
+                }
             }
         } catch (SQLException e) {
             throw new RuntimeException("Error inserting announcement into database", e);
@@ -57,9 +64,9 @@ public class AnnouncementsDao {
             PreparedStatement ps = c.prepareStatement(sql);
              ResultSet rs = ps.executeQuery()){
 
-            if (rs.next())
+            while(rs.next()) {
                 announcements.add(retrieveAnnouncement(rs));
-
+            }
         } catch (SQLException e) {
             throw new RuntimeException("Error querying announcements from database", e);
         }
@@ -68,6 +75,9 @@ public class AnnouncementsDao {
     }
 
     public void removeAnnouncement(long announcementId){
+        if(!contains(announcementId)){
+            return;
+        }
         String sql = "DELETE FROM announcements WHERE announcement_id = ?";
 
         try (Connection c = dataSource.getConnection();
@@ -76,11 +86,60 @@ public class AnnouncementsDao {
             ps.setLong(1, announcementId);
 
             ps.executeUpdate();
+            cnt--;
         } catch (SQLException e) {
             throw new RuntimeException("Error removing announcement from database", e);
         }
     }
+    public boolean contains(Announcement a){
+        if(a==null){
+            return false;
+        }
+        String sql = "SELECT COUNT(*) FROM announcements WHERE announcement_id=?" +
+                "AND administrator_id=? AND announcement_text=? AND creation_date=?";
 
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+
+            ps.setLong(1, a.getAnnouncementId());
+            ps.setLong(2,a.getAdministratorId());
+            ps.setString(3, a.getAnnouncementText());
+            ps.setTimestamp(4,a.getCreationDate());
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+    public boolean contains(long aid){
+        if(aid<0||aid>cnt){
+            return false;
+        }
+        String sql = "SELECT COUNT(*) FROM announcements WHERE announcement_id=?";
+
+        try (Connection connection = dataSource.getConnection();
+             PreparedStatement ps = connection.prepareStatement(sql)) {
+
+            ps.setLong(1, aid);
+            ResultSet rs = ps.executeQuery();
+
+            if (rs.next()) {
+                return rs.getInt(1) > 0;
+            }
+
+            return false;
+        } catch (SQLException e) {
+            throw new RuntimeException("Error checking user existence", e);
+        }
+    }
+    public long getCount(){
+        return cnt;
+    }
     private Announcement retrieveAnnouncement(ResultSet rs) throws SQLException {
         return new Announcement(rs.getLong("announcement_id"), rs.getLong("administrator_id"),
                 rs.getString("announcement_text"), rs.getTimestamp("creation_date"));
